Revert "Refactor Visualization and Update Student Notebook"
