+++ docs/conf.py
@@ -20,7 +20,6 @@
 import os
 import sys
 sys.path.append(os.path.abspath('../'))
-sys.path.insert(0, os.path.abspath('../gos'))
 
 # -- General configuration ------------------------------------------------
 
@@ -50,7 +49,7 @@ master_doc = 'index'
 
 # General information about the project.
 project = 'GOS'
-copyright = '2019, Center for Research Computing'
+copyright = '2017, Carl Colglazier'
 author = 'Carl Colglazier'
 
 # The version info for the project you're documenting, acts as replacement for
