+++ docs/index.rst
@@ -1,9 +1,17 @@
 .. include:: ../README.rst
 
+Examples
+--------
+             
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   source/*
+
+
 API
 ---
 
 .. automodule:: gos
-                :members:
-.. automodule:: visualization
                 :members:
