+++ gos/World/World.cpg
