+++ gos/__init__.py
@@ -1,4 +1,5 @@
 from .gos import *
+from .visulize import *
 
 __all__ = ["World", "Globe", "Agent"]
 __title__ = 'globalopensimulator'
