--- gos/visualization/__init__.py
@@ -1,8 +1,0 @@
-from .map import map_plot, matrix_plot
-from .midpointnorm import MidPointNorm
-import pandas as pd
-
-COLOR1 = '#46bcec'
-COLOR2 = '#eeeeee'
-COLOR3 = '#333333'
-COLOR4 = '#000000'
