--- gos/visualization/midpointnorm.py
@@ -1,70 +1,0 @@
-from matplotlib.colors import Normalize
-from numpy import ma
-from matplotlib import cbook
-
-# https://stackoverflow.com/questions/7404116/defining-the-midpoint-of-a-colormap-in-matplotlib
-class MidPointNorm(Normalize):
-    """
-    Helper normalization class.
-    """
-    def __init__(self, midpoint=0, vmin=None, vmax=None, clip=False):
-        Normalize.__init__(self,vmin, vmax, clip)
-        self.midpoint = midpoint
-
-    def __call__(self, value, clip=None):
-        if clip is None:
-            clip = self.clip
-
-        result, is_scalar = self.process_value(value)
-
-        self.autoscale_None(result)
-        vmin, vmax, midpoint = self.vmin, self.vmax, self.midpoint
-
-        if not (vmin < midpoint < vmax):
-            raise ValueError("midpoint must be between maxvalue and minvalue.")       
-        elif vmin == vmax:
-            result.fill(0) # Or should it be all masked? Or 0.5?
-        elif vmin > vmax:
-            raise ValueError("maxvalue must be bigger than minvalue")
-        else:
-            vmin = float(vmin)
-            vmax = float(vmax)
-            if clip:
-                mask = ma.getmask(result)
-                result = ma.array(np.clip(result.filled(vmax), vmin, vmax),
-                                  mask=mask)
-
-            # ma division is very slow; we can take a shortcut
-            resdat = result.data
-
-            #First scale to -1 to 1 range, than to from 0 to 1.
-            resdat -= midpoint            
-            resdat[resdat>0] /= abs(vmax - midpoint)            
-            resdat[resdat<0] /= abs(vmin - midpoint)
-
-            resdat /= 2.
-            resdat += 0.5
-            result = ma.array(resdat, mask=result.mask, copy=False)                
-
-        if is_scalar:
-            result = result[0]            
-        return result
-
-    def inverse(self, value):
-        if not self.scaled():
-            raise ValueError("Not invertible until scaled")
-        vmin, vmax, midpoint = self.vmin, self.vmax, self.midpoint
-
-        if cbook.iterable(value):
-            val = ma.asarray(value)
-            val = 2 * (val-0.5)  
-            val[val>0]  *= abs(vmax - midpoint)
-            val[val<0] *= abs(vmin - midpoint)
-            val += midpoint
-            return val
-        else:
-            val = 2 * (val - 0.5)
-            if val < 0: 
-                return  val*abs(vmin-midpoint) + midpoint
-            else:
-                return  val*abs(vmax-midpoint) + midpoint
