+++ requirements.txt
@@ -1,4 +1,3 @@
-numpy>=1.12.1
-pandas>=0.20.1
-sphinxcontrib-bibtex>=0.3.6
-Cartopy>=0.17.0
+numpy==1.12.1
+pandas==0.20.1
+sphinxcontrib-bibtex==0.3.6
